Fixed many -Wshorten-64-to-32 truncation warnings; no public API changes
In most cases I increased some variable sizes.

In some cases I added a cast, where the value was guaranteed to fit in the smaller destination variable size.

Changed one atoll() to strtoll() since it does better error checking. Check for errors and bail if the value does not fit into an `int`.

No public API have changed.

+++ nifti2/clib_02_nifti2.c
@@ -117,7 +117,7 @@ int disp_float(nifti_image * nim, int vol, int slice, int line, int offset)
 {
    float     * dp, * d2;
    long long   lloff;
-   int         nx, nxy, nxyz;
+   int64_t     nx, nxy, nxyz;
 
    if( ! nim ) return 1;
